+++ packages/orbit-components/src/Modal/ModalHeader/index.d.ts
@@ -1,8 +1,7 @@
 // Type definitions for @kiwicom/orbit-components
 // Project: http://github.com/kiwicom/orbit
-import * as React from "react";
+import React from "react";
 
-import { Props as IllustrationProps } from "../../Illustration";
 import * as Common from "../../common/common";
 
 declare module "@kiwicom/orbit-components/lib/Modal/ModalHeader";